fix bug 🐛 when guess prompt is cancelled
previously when the prompt of the guess was cancelled the game did
did not stop.

I solved this by splitting the expression that converts the prompt
result to a number into two expressions, one that retrieves the prompt
and another expression that converts it to a number. Between the two
expressions, I was able to check for a "null" which indicated that
the prompt was cancelled by the user.

I also improved the code by using two variables for the prompt before
the numeric conversion and after so that it is compatible with
TypeScript.